+++ src/killstreak_tier.rs
@@ -1,4 +1,4 @@
-use crate::Attribute;
+use crate::{Attribute, EffectType, DescriptionFormat};
 use strum_macros::{Display, EnumString, EnumIter, EnumCount};
 use num_enum::{TryFromPrimitive, IntoPrimitive};
 use serde_repr::{Serialize_repr, Deserialize_repr};
@@ -38,8 +38,8 @@ impl Attribute for KillstreakTier {
     const NAME: &str = "killstreak tier";
     const ATTRIBUTE_CLASS: &str = "killstreak_tier";
     const DESCRIPTION_STRING: Option<&str> = Some("Killstreaks Active");
-    const DESCRIPTION_FORMAT: Option<&str> = Some("value_is_additive");
-    const EFFECT_TYPE: &str = "positive";
+    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsAdditive);
+    const EFFECT_TYPE: EffectType = EffectType::Positive;
     const HIDDEN: bool = false;
     const STORED_AS_INTEGER: bool = false;
 }