+++ src/sheen.rs
@@ -1,4 +1,4 @@
-use crate::Attribute;
+use crate::{Attribute, EffectType, DescriptionFormat};
 use strum_macros::{Display, EnumString, EnumIter, EnumCount};
 use num_enum::{TryFromPrimitive, IntoPrimitive};
 use serde_repr::{Serialize_repr, Deserialize_repr};
@@ -46,8 +46,8 @@ impl Attribute for Sheen {
     const NAME: &str = "killstreak idleeffect";
     const ATTRIBUTE_CLASS: &str = "killstreak_idleeffect";
     const DESCRIPTION_STRING: Option<&str> = Some("Sheen: %s1");
-    const DESCRIPTION_FORMAT: Option<&str> = Some("value_is_killstreak_idleeffect_index");
-    const EFFECT_TYPE: &str = "positive";
+    const DESCRIPTION_FORMAT: Option<DescriptionFormat> = Some(DescriptionFormat::ValueIsKillstreakIdleEffectIndex);
+    const EFFECT_TYPE: EffectType = EffectType::Positive;
     const HIDDEN: bool = false;
     const STORED_AS_INTEGER: bool = false;
 }